Improve memory consumption
For every established GTP-U connection, several metric probes are
created upon connection (see src/gtp_u_edp_port.erl). Among the created
probes are 2 of type "histogram". An histogram is represented as a
process in the GTP-U-Node which will be initialized with a
`min_heap_size` of 40k WORDS (8byte on 64bit machines).

See [1] and [2] for a discussion about the issue.

Neither of the preconditions for increasing the `min_heap_size` are met:

* the histograms are kept over a long period of time (and are not short
  lived)
* the amount of processes keep growing with the amount of established
  connections

This commit sets the `min_heap_size` option of the spawned processes
responsible for the histogram to the default Erlang value and thus,
decreases the memory consumption of the histograms and allow the GC to
work.

[1]: https://github.com/Feuerlabs/exometer_core/issues/100
[2]: http://erlang.org/doc/efficiency_guide/processes.html